Corrects projects event listener
